Add name to workspace reference
